refactor(desktop): split session runtime dev SSE logging
