+++ apps/desktop/src/lib/infra/debug/dev-sse-event-record.ts
@@ -1,0 +1,10 @@
+import type { SessionEventEnvelope } from "@anyharness/sdk";
+
+export type DevSSEEventStatus = "applied" | "duplicate" | "gap";
+
+export type DevSSEEventRecord = {
+  sessionId: string;
+  receivedAt: string;
+  status: DevSSEEventStatus;
+  envelope: SessionEventEnvelope;
+};
