+++ apps/desktop/src/lib/infra/debug/dev-transcript-phase-log.ts
@@ -1,11 +1,5 @@
-import type { ContentPart, SessionEventEnvelope } from "@anyharness/sdk";
-
-type DevSSEEventRecord = {
-  sessionId: string;
-  receivedAt: string;
-  status: "applied" | "duplicate" | "gap";
-  envelope: SessionEventEnvelope;
-};
+import type { SessionEventEnvelope } from "@anyharness/sdk";
+import type { DevSSEEventStatus } from "@/lib/infra/debug/dev-sse-event-record";
 
 type DevTranscriptPhaseKind =
   | "turn"
@@ -19,7 +13,7 @@ type DevTranscriptPhaseKind =
 type DevTranscriptPhaseRecord = {
   sessionId: string;
   recordedAt: string;
-  status: DevSSEEventRecord["status"];
+  status: DevSSEEventStatus;
   seq: number;
   eventTimestamp: string;
   eventType: string;
@@ -56,139 +50,10 @@ type DevTranscriptPhaseSessionState = {
   itemsById: Record<string, DevTranscriptPhaseItemState>;
 };
 
-type DevSessionRuntimeRecord = {
-  sessionId: string;
-  recordedAt: string;
-  kind: string;
-  details: Record<string, unknown>;
-};
-
-export function logDevSSEEvent(
+export function logDevTranscriptPhaseEvent(
   sessionId: string,
   envelope: SessionEventEnvelope,
-  status: DevSSEEventRecord["status"],
-): void {
-  if (!import.meta.env.DEV) {
-    return;
-  }
-
-  const debugGlobal = globalThis as typeof globalThis & {
-    __APOLLO_SSE_EVENTS__?: DevSSEEventRecord[];
-  };
-
-  const record: DevSSEEventRecord = {
-    sessionId,
-    receivedAt: new Date().toISOString(),
-    status,
-    envelope: sanitizeDevSSEEnvelope(envelope),
-  };
-
-  const existing = debugGlobal.__APOLLO_SSE_EVENTS__ ?? [];
-  debugGlobal.__APOLLO_SSE_EVENTS__ = [...existing.slice(-499), record];
-  logDevTranscriptPhaseEvent(sessionId, envelope, status);
-}
-
-export function logDevSessionRuntimeEvent(
-  sessionId: string,
-  kind: string,
-  details: Record<string, unknown>,
-): void {
-  if (!import.meta.env.DEV) {
-    return;
-  }
-
-  const debugGlobal = globalThis as typeof globalThis & {
-    __APOLLO_SESSION_RUNTIME__?: DevSessionRuntimeRecord[];
-    __APOLLO_SESSION_RUNTIME_CONSOLE__?: boolean;
-  };
-
-  const record: DevSessionRuntimeRecord = {
-    sessionId,
-    recordedAt: new Date().toISOString(),
-    kind,
-    details,
-  };
-
-  const existing = debugGlobal.__APOLLO_SESSION_RUNTIME__ ?? [];
-  debugGlobal.__APOLLO_SESSION_RUNTIME__ = [...existing.slice(-499), record];
-  if (debugGlobal.__APOLLO_SESSION_RUNTIME_CONSOLE__ === true) {
-    console.debug("[session-runtime]", kind, record);
-  }
-}
-
-function sanitizeDevSSEEnvelope(envelope: SessionEventEnvelope): SessionEventEnvelope {
-  const event = envelope.event;
-  if (event.type === "item_started" || event.type === "item_completed") {
-    return {
-      ...envelope,
-      event: {
-        ...event,
-        item: {
-          ...event.item,
-          contentParts: sanitizeContentParts(event.item.contentParts ?? []),
-          rawInput: undefined,
-          rawOutput: undefined,
-        },
-      },
-    };
-  }
-  if (event.type === "item_delta") {
-    return {
-      ...envelope,
-      event: {
-        ...event,
-        delta: {
-          ...event.delta,
-          replaceContentParts: event.delta.replaceContentParts
-            ? sanitizeContentParts(event.delta.replaceContentParts)
-            : undefined,
-          appendContentParts: event.delta.appendContentParts
-            ? sanitizeContentParts(event.delta.appendContentParts)
-            : undefined,
-          rawInput: undefined,
-          rawOutput: undefined,
-        },
-      },
-    };
-  }
-  if (event.type === "pending_prompt_added" || event.type === "pending_prompt_updated") {
-    return {
-      ...envelope,
-      event: {
-        ...event,
-        text: summarizeSanitizedContent(event.contentParts ?? [], event.text),
-        contentParts: sanitizeContentParts(event.contentParts ?? []),
-      },
-    };
-  }
-  return envelope;
-}
-
-function sanitizeContentParts(parts: ContentPart[]): ContentPart[] {
-  return parts.map((part) => {
-    switch (part.type) {
-      case "text":
-        return { type: "text", text: `[text:${part.text.length}]` };
-      case "resource":
-        return { ...part, preview: part.preview ? `[preview:${part.preview.length}]` : undefined };
-      case "tool_input_text":
-        return { type: "tool_input_text", text: `[text:${part.text.length}]` };
-      case "tool_result_text":
-        return { type: "tool_result_text", text: `[text:${part.text.length}]` };
-      default:
-        return part;
-    }
-  });
-}
-
-function summarizeSanitizedContent(parts: ContentPart[], fallback: string): string {
-  return parts.length > 0 ? `[content_parts:${parts.length}]` : `[text:${fallback.length}]`;
-}
-
-function logDevTranscriptPhaseEvent(
-  sessionId: string,
-  envelope: SessionEventEnvelope,
-  status: DevSSEEventRecord["status"],
+  status: DevSSEEventStatus,
 ): void {
   if (!import.meta.env.DEV) {
     return;
